output: fold syslog_tls into syslog_tcp as per-peer TLS
Remove the standalone `syslog_tls` output module. `syslog_tcp` now
takes a per-peer `tls` block (inline or named-profile reference) so
plaintext and TLS destinations can share one peer list.

Internals: a `Conn { Plain(TcpStream) | Tls(Box<TlsStream<...>>) }`
enum with delegated AsyncWrite implements both shapes behind the
existing `write_framed`; `PeerList<Conn>` and rotation/cooldown
semantics are unchanged. TLS connectors are built up-front into
`Vec<Option<TlsConnector>>` indexed by peer; the crypto provider is
installed only when at least one peer is TLS.

Default port is per-peer: 6514 (RFC 5425) when `tls` is set on that
peer, 514 (RFC 6587) otherwise. Framing remains output-wide.

BREAKING: `type syslog_tls` outputs must be renamed to `type
syslog_tcp`. Existing `tls { profile { ... } }` and per-peer
`tls { ... }` / `tls <profile>` shapes work unchanged.

The `input syslog_tls` listener is untouched.
