+++ crates/limpid/src/modules/output/mod.rs
@@ -9,6 +9,5 @@ pub(crate) mod persistent_conn;
 pub mod stdout;
 pub(crate) mod syslog_peers;
 pub mod syslog_tcp;
-pub mod syslog_tls;
 pub mod syslog_udp;
 pub mod unix_socket;
