+++ docs/src/outputs/README.md
@@ -9,9 +9,8 @@ Output modules write processed events to external destinations.
 | [`file`](./file.md) | Local file with dynamic path templates |
 | [`http`](./http.md) | HTTP/HTTPS endpoint (Elasticsearch, Splunk HEC, etc.) |
 | [`kafka`](./kafka.md) | Apache Kafka topic (requires `--features kafka`) |
-| [`syslog_tcp`](./syslog-tcp.md) | Syslog TCP with persistent per-peer connections |
+| [`syslog_tcp`](./syslog-tcp.md) | Syslog TCP with persistent per-peer connections; per-peer TLS optional |
 | [`syslog_udp`](./syslog-udp.md) | Syslog UDP datagrams |
-| [`syslog_tls`](./syslog-tls.md) | Syslog TCP over TLS with optional client certificate |
 | [`unix_socket`](./unix-socket.md) | Unix stream socket |
 | [`stdout`](./stdout.md) | Standard output (debugging) |
 | [`otlp`](./otlp.md) | OTLP logs sender (HTTP/JSON, HTTP/protobuf, gRPC) |
